Delete docs/flutter directory
Removes the directory because right now the structure is incongruous to the goals:

Currently the directory is (and always has been) copying "flutterfire/docs" into "flutter/". Notice how this was one directory displaced from where it "should be". With the upstream change "flutterfire/docs" will map to "flutter/docs". Again there is a difference (the absence of Fire) but this is in keeping with the previous structure and nomenclature.